fix: sole connectWallet
