+++ src/pages/bridge/Header/index.tsx
@@ -1,20 +1,14 @@
 import { Button, Container } from "@mui/material"
 
 import { useWeb3Context } from "@/contexts/Web3ContextProvider"
-import { goMetaMaskApp } from "@/utils"
+import useConnectWallet from "@/hooks/useConnectWallet"
 
 import AddressButton from "./AddressButton"
 
 const Header = () => {
-  const { walletCurrentAddress, connectWallet } = useWeb3Context()
+  const { walletCurrentAddress } = useWeb3Context()
+  const connectWallet = useConnectWallet()
 
-  const handleConnectWallet = () => {
-    if (typeof window.ethereum !== "undefined") {
-      connectWallet()
-    } else {
-      goMetaMaskApp()
-    }
-  }
   return (
     <Container
       sx={{
@@ -30,7 +24,7 @@ const Header = () => {
           <AddressButton />
         </>
       ) : (
-        <Button sx={{ width: "17.8rem" }} onClick={handleConnectWallet} variant="outlined">
+        <Button sx={{ width: "17.8rem" }} onClick={connectWallet} variant="outlined">
           Connect Wallet
         </Button>
       )}
