+++ src/pages/rollup/components/Header.tsx
@@ -2,7 +2,8 @@ import { Box, Button, Typography } from "@mui/material"
 import { styled } from "@mui/material/styles"
 
 import { useWeb3Context } from "@/contexts/Web3ContextProvider"
-import { goMetaMaskApp, truncateAddress } from "@/utils"
+import useConnectWallet from "@/hooks/useConnectWallet"
+import { truncateAddress } from "@/utils"
 
 const Description = styled(Typography)(({ theme }) => ({
   maxWidth: "61rem",
@@ -19,21 +20,15 @@ const AccountBox = styled(Box)(({ theme }) => ({
 }))
 
 const Header = () => {
-  const { walletCurrentAddress, connectWallet } = useWeb3Context()
-  const handleConnectWallet = () => {
-    if (typeof window.ethereum !== "undefined") {
-      connectWallet()
-    } else {
-      goMetaMaskApp()
-    }
-  }
+  const { walletCurrentAddress } = useWeb3Context()
+  const connectWallet = useConnectWallet()
   return (
     <Box className="mx-auto">
       <AccountBox>
         {walletCurrentAddress ? (
           <Button className="w-[178px] pointer-events-none">{truncateAddress(walletCurrentAddress as string)}</Button>
         ) : (
-          <Button onClick={handleConnectWallet}>Connect Wallet</Button>
+          <Button onClick={connectWallet}>Connect Wallet</Button>
         )}
       </AccountBox>
       <Typography variant="h3" align="center" sx={{ marginBottom: "1.2rem" }}>
